requirements.txt
tensorflow>=2.8.0
tensorflow-addons>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=8.3.0
tqdm>=4.62.0
opencv-python>=4.5.0
wandb>=0.12.0

